requirements.txt
argparse==1.2.1
beautifulsoup4==4.3.2
requests==2.5.0
wsgiref==0.1.2
MySQL-python==1.2.5
statsd-client==1.0.4
